backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.5
pymongo==4.6.1
motor==3.3.2
Pillow==9.5.0
python-dotenv==0.19.0
requests==2.31.0
pydantic==2.6.3
geopy==2.4.1
opencage==2.3.1
aiofiles==23.2.1